requirements.txt
pandas
scikit-learn
streamlit
joblib
spacy
google-generativeai
openai
textblob
transformers
datasets
